requirements.txt
streamlit==1.29.0
langchain==0.1.0
langchain-openai==0.0.2
openai==1.6.1
python-dotenv==1.0.0
pydantic==2.5.2
jsonschema==4.20.0
pandas==2.1.4
plotly==5.17.0
pytest==7.4.3 